challenge/requirements.txt
# Python==3.10.9
# pip==22.3.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1